+++ components/guide/GuideBaseButtons.vue
@@ -217,6 +217,17 @@ export default {
             }
           },
           {
+            name: "loading",
+            type: {
+              type: "Boolean",
+              additional: "Default: false"
+            },
+            details: {
+              description:
+                "Set the button as loading and display a small spinner inside it."
+            }
+          },
+          {
             name: "reverse",
             type: {
               type: "Boolean",